pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "rgc-llm"
version = "0.2.0"
description = "RGC-LLM engineering prototype: resonant graph calculus language model"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "torch>=2.2",
  "pydantic>=2.6",
  "PyYAML>=6.0",
  "typer>=0.12",
  "rich>=13.7",
  "numpy>=1.26"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.0"
]

[project.scripts]
rgc-train = "rgc_llm.cli.train:app"
rgc-infer = "rgc_llm.cli.infer:app"
rgc-export = "rgc_llm.cli.export:app"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
